+++ dspublisher/theme/styles/banners.css
@@ -1,8 +1,37 @@
 .toc-banner-source {
-    position: fixed;
-    top: 0;
-    left: 0;
-    visibility: hidden;
-    height: 3px;
-    z-index: -9999;
+  position: fixed;
+  top: 0;
+  left: 0;
+  visibility: hidden;
+  height: 3px;
+  z-index: -9999;
+}
+
+.toc-banner {
+  margin-top: 2rem;
+  position: relative;
+  width: 11.25rem;
+}
+
+.toc-banner > img {
+  width: 100%;
+  height: auto;
+}
+
+.toc-banner--content {
+  margin-top: 1rem;
+  line-height: 1.5;
+}
+
+.toc-banner > a {
+  display: block;
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+}
+
+.toc-banner > a::after {
+  content: none !important;
 }
